Refactor of functions
